requirements.txt
python-telegram-bot==20.0
Pillow
python-dotenv
pydub==0.25.1
ffmpeg-python==0.2.0
numpy==1.21.0
